# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for using Swagger with Echo.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/clients": {
            "get": {
                "description": "Recupera una lista de todos los clientes en la base de datos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtiene todos los clientes",
                "responses": {
                    "200": {
                        "description": "Lista de clientes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Client"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo cliente con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear cliente",
                "parameters": [
                    {
                        "description": "Información del Cliente",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                }
            }
        },
...
        "/api/v1/clients/{id}": {
            "get": {
                "description": "Recupera un cliente específico usando su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener cliente por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del cliente",
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza la información de un cliente existente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Información actualizada del Cliente",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un cliente específico usando su ID",
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente eliminado exitosamente"
                    }
                }
            }
        },
...
        "/groups": {
            "get": {
                "description": "Recupera todos los grupos de la base de datos",
                "tags": [
                    "Group"
                ],
                "summary": "Obtiene todos los grupos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Group"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo grupo en la base de datos",
                "tags": [
                    "Group"
                ],
                "summary": "Crea un grupo",
                "parameters": [
                    {
                        "description": "Datos del grupo",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                }
            }
        },
        "/groups/{group_id}": {
            "delete": {
                "description": "Elimina un grupo de la base de datos basado en el ID proporcionado",
                "tags": [
                    "Group"
                ],
                "summary": "Elimina un grupo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del grupo",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Group deleted"
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "description": "Recupera un grupo específico basado en el ID proporcionado",
                "tags": [
                    "Group"
                ],
                "summary": "Obtiene un grupo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del grupo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/groups/{group_id}": {
            "put": {
                "description": "Asigna un grupo existente a un usuario basado en el ID del usuario y del grupo",
                "tags": [
                    "Group"
                ],
                "summary": "Asigna un grupo a un usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del grupo",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group assigned"
                    }
                }
            },
            "delete": {
                "description": "Elimina la relación de un grupo asignado a un usuario basado en sus IDs",
                "tags": [
                    "Group"
                ],
                "summary": "Elimina la asignación de un grupo a un usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del grupo",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group unassigned"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Client": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Group": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}